feat: add ability to pull all subtrees standalone.
