+++ scripts/helpers/openSourceProjects.js
@@ -1,0 +1,4 @@
+module.exports = new Set([
+  '@haiku/core',
+  '@haiku/cli'
+])
